requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12

# LangChain and LLM
langchain==0.3.7
langchain-core==0.3.15
langchain-google-genai==2.0.5
google-generativeai==0.8.3

# Vector store and embeddings
faiss-cpu==1.9.0
FlagEmbedding==1.3.2

# NLP and retrieval
rank-bm25==0.2.2
nltk==3.9.1
sentence-transformers==3.3.1

# ML/AI frameworks
torch==2.5.1
transformers==4.44.2  # Required by FlagEmbedding 1.3.2

# Utilities
numpy==1.26.4
python-dotenv==1.0.1
